refactor(auth): extract token generation to internal package
